Remove redundant "phpcs" command from workflow
The "command: phpcs" line was removed from the phpcs.yml workflow file. This change was necessary as the command is already specified in the php-coding-standards.yml workflow from alleyinteractive repository and doesn’t need to be defined twice. This refactoring helps to avoid confusion and potential command conflicts in the future.