Ignore specific PHP Code Sniffer warnings for audio node checks
This change was necessary to ensure the codebase conforms to the WordPress Naming Conventions. Comments were added to ignore PHP Code Sniffer warnings about non-snake case variables when checking if a node is an 'audio' node. This had to be done since those checks are part of the application's logic