"Add PHP Code Sniffer ignores for snake case naming checks
The project has been updated to add comments to ignore the 'not using snake case' warnings from PHP Code Sniffer. This was necessary as the property names used in the Node class provided by the core PHP library, do not follow snake case convention. These warnings were leading to unnecessary noise in the linting reports and were misleading as it could lead developers to think that there's something wrong with the project's own codebase. As this convention is not something developer can change, it is now simply ignored by the linter."