Change submenu parent for Bulk Export page to handle PHP deprecation notices.
In this commit, the Bulk Export submenu in the apple-news plugin was changed to not appear in any menu. The parent of the submenu was previously null, but it has now been set to 'options.php'. This change is intended to provide more organization and decrease clutter in the main menu.