Update version requirements and contributors in readme
The readme file has been updated to reflect the new minimum requirements for the software: WordPress version was raised to 5.2, PHP version is now 7.4. A new contributor was added to the list and a revision in the updates history, specifically on version 2.0.0, was made. These changes are primarily to maintain the accuracy and relevance of the information in the readme file.