Remove temporary PHPUnit Testing Suite, adjust composer.json and modify bootstrap.php for tests
Deleted the temporary .github/workflows/php-tests-temp.yml file as it is no longer needed. Adjusted composer.json to eliminate a duplicated plugin. In tests/bootstrap.php, moved some code blocks inside the function and also set the WP_TESTS_MULTISITE constant to 1. This cleanup and reorganization improves the readability and efficiency of code and aligns with testing requirements.