Update GitHub Action to abort on non-incremented versions
The GitHub Action workflow which deploys to WordPress is updated to exit with an error if the version is not incremented compared to the previous one. This change is introduced to prevent unintentional overwrites of existing plugin versions on WordPress. New 'Build' and 'Publish' steps will run only when version increment is detected, ensuring that each published version is unique.