Update GitHub Actions workflow for better efficiency
This commit moves the GITHUB_TOKEN to the environment section of the release job and removes it from individual steps for better readability. It also improves the version comparison condition for more accuracy, and adds condition checks for each release-related task to only run if the version has incremented. This reduces unnecessary job executions, leading to quicker and more efficient workflow runs.