Add release instructions to README
This commit adds a new section to the README file, detailing the steps required to release the plugin. It includes information on how the plugin utilises a built release workflow to compile and tag releases, as well as instructions for initiating a new release. This addition aims to provide clarity for future contributors or maintainers of the project.