PHPCS fixes
This commit carries out a number of improvements to enhance both the readability and functionality. Firstly, it fixes a comment punctuation issue in 'class-apple-news.php'. Secondly, it substitutes 'json_encode()' with 'wp_json_encode()' in 'class-request.php' to ensure the proper encoding of data in WordPress. Finally, it cleans up the formatting in 'class-export.php', removing unnecessary backslashes and adding parentheses for instantiation of 'BC_Setup' class.