Update condition check and get method in settings to fix bundle URL bug
The code is updated to check if the filename is `null` instead of an empty string. This is a more accurate way to verify if filename is set before bundling source. Additionally, the way settings are referred to has been changed. Now the magic method `__get` is used instead of the `get` method, to ensure proper retrieval of the property regardless the visibility of the property.