Update node, npm version specifications in Apple News plugin
The node and npm version specifications in package.json and package-lock.json files of the Apple News plugin are modified. Instead of specifying a fixed range, a more flexible "^20.x.x" for node and "^10.x.x" for npm are now used, indicating any minor version updates within the major versions 20 and 10 respectively are acceptable.